// src/types/generated/GetImagesPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TimelineSort } from "./TimelineSort";

/**
 * タイムライン画像一覧取得リクエストペイロード
 */
export type GetImagesPayload = { 
/**
 * 取得開始オフセット
 */
offset: number, 
/**
 * 取得上限数（最大500）
 */
limit: number, 
/**
 * 特定フォルダによる絞り込み（オプション）
 */
folderId: number | null, 
/**
 * 表示ソート順（オプション、デフォルトはTakenAtDesc）
 */
sort: TimelineSort | null, };